Reset version number to 2.0.0-SNAPSHOT
Reset dependency versions to SNAPSHOT versions.

Usually, after a release we would increment the version number (e.g.,
after release 2.0.0 we would go to 2.0.1-SNAPSHOT) but for the betas,
we return to 2.0.0-SNAPSHOT again.

This one time, we are changing the version number from 2.0-SNAPSHOT to
2.0.0-SNAPSHOT, adding the extra zero. Ideally we would have had the
major.minor.micro pattern this whole time, but better late than never.
Now is as good a time as any to start.