RealPositionable Local neighborhoods: solve Dissociative identity disorder
The calibrated center coordinates are given by the AbstractNeighborhood.center
field. The size of the neighborhood is set in pixel units by the span long[]
array. Which leads us to:

Dissociative identity disorder: This class is a RealPositionable and thus a
Positionable. But the two groups of methods have different meanings:

All the move(long[]) and co. methods that operate on int or long arrays or on
Localizable operate on the pixel position of the neighborhood. So you can still
move pixel by pixel.

All the move(double[]) and co. methods that operate on double or float arrays
or on RealLocalizable operates on the calibrated position of the neighborhood,
owing to the calibration.  An example: Suppose you have an ImgPlus spanning 100
x 100 pixels, with a pixel size of 0.2 µm. If you use setPosition(50L, 0) you
will move the center of this neighborhood to the center of the image. But using
setPosition(50f, 0) will move you to the position x = 50 µm, which is out of
bounds. So be careful.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>