Remove duplicate Neighborhood interface
The exact same class
(net.imglib2.algorithm.region.localneighborhood.Neighborhood) lives in
the imglib2-algorithms project as well.

Even though this version (in imglib2-algorithms-gpl) came first
according to the Git history, it seems it is not used anywhere, whereas
the version in imglib2-algorithms is used in several places.

So to avoid conflicts, this version has been removed. If anything needs
it, there is no problem to reinstate it as long as the name is changed.