POM: update to standalone pom-imglib2 parent
We no longer need to declare most versions, since they are managed.

The artifact was renamed to imglib2-algorithm-gpl (from
imglib2-algorithms-gpl) for consistency with its package prefix.

(Even though the prefix is net.imglib2.algorithm rather than
net.imglib2.algorithm.gpl, the gpl suffix is not indicative of
subpackage functionality in this case, but rather a licensing
distinction, and thus is not appended to the package prefix, as
discussed at the October 2014 Madison ImgLib2 hackathon.)

Because of the artifactId change, we can start the versioning
over at 0.1.0, which according to SemVer indicates the initial
incubation period of development.

Lastly, this change makes the necessary code changes to adjust
for the backwards incompatibilities of the ImgLib2 release.